+++ src/main/java/ar/edu/utn/frc/tup/lc/iv/dtos/get/GetOwnerUserDto.java
@@ -97,5 +97,5 @@ public class GetOwnerUserDto {
     /**
      * Identificador de la plataforma telegram utilizada en notificaciones.
      */
-    private Integer telegram_id;
+    private Long telegram_id;
 }
