+++ src/main/java/ar/edu/utn/frc/tup/lc/iv/dtos/post/BasePostUser.java
@@ -99,7 +99,7 @@ public class BasePostUser {
     /**
      * Identificador de la plataforma telegram utilizada en notificaciones.
      */
-    private Integer telegram_id;
+    private Long telegram_id;
 
     /**
      * Constructor protegido para evitar instanciar la clase directamente.
